Fix bugs and allow multiple sync services
