+++ src/models/ienv.model.ts
@@ -1,0 +1,5 @@
+export interface IEnv {
+  [key: string]: string | undefined;
+  http_proxy: string;
+  HTTP_PROXY: string;
+}
