+++ src/models/ifixgistresponse.model.ts
@@ -1,0 +1,6 @@
+import * as GitHubApi from "@octokit/rest";
+
+export interface IFixGistResponse
+  extends Omit<GitHubApi.GistsGetResponse, "files"> {
+  files: any | GitHubApi.GistsGetResponseFiles;
+}
