+++ src/models/sync.model.ts
@@ -1,0 +1,6 @@
+export interface ISyncService {
+  id: string;
+  UploadSettings(): Promise<void>;
+  DownloadSettings(): Promise<void>;
+  Connect(): Promise<void>;
+}
